+++ src/app/features/pagos/models/pago.model.ts
@@ -7,10 +7,7 @@ export interface Pago {
   vr_devengado: number;
   vr_deduccion: number;
   vr_neto: number;
-}
-
-export interface PagosQueryParams {
-  page?: number;
-  size?: number;
-  empleado_id: number;
+  codigo_pago_tipo_fk: string;
+  pago_tipo_nombre: string;
+  numero: number;
 }